Get tests working. Update lifecycle tests to reflect native behaviour. Firefox now breaks.
