+++ src/util/ready.js
@@ -1,0 +1,9 @@
+export default function (fn) {
+  return function () {
+    if (document.readyState === 'complete') {
+      fn();
+    } else {
+      document.addEventListener('DOMContentLoaded', fn);
+    }
+  };
+}
