+++ test/unit/api/property.js
@@ -116,111 +116,11 @@ describe('api/property', function () {
     expect(el.hasAttribute('prop-name1')).to.equal(false, 'Attribute should be removed if value is false');
   });
 
-  it('events', function () {
-    var triggered = 0;
-
-    skate(elem.safe, {
-      properties: {
-        propName1: null
-      }
-    });
-
-    var el = elem.create();
-    el.addEventListener('skate.property', () => ++triggered);
-    el.propName1 = true;
-    expect(triggered).to.equal(1);
-  });
-
-  it('dependencies', function () {
-    var triggered = 0;
-
-    skate(elem.safe, {
-      properties: {
-        propName1: null,
-        propName2: {
-          deps: 'propName1',
-          set: function () {
-            ++triggered;
-          }
-        }
-      }
-    });
-
-    var el = elem.create();
-
-    // Triggered twice. Once for propName1 and once for propName2.
-    el.addEventListener('skate.property', () => ++triggered);
-
-    // Will notify propName2.
-    el.propName1 = true;
-
-    // Twice for `skate.property` and once for the `propName2` setter.
-    expect(triggered).to.equal(3);
-  });
-
-  it('dependencies (deep; elements must be in DOM)', function () {
-    var triggeredNumber = 0;
-    var triggeredTarget;
-    var elem2 = helperElement();
-
-    skate(elem.safe, {
-      properties: {
-        propName1: {
-          deps: `propName1 > *`,
-          set: () => ++triggeredNumber
-        }
-      }
-    });
-
-    skate(elem2.safe, {
-      properties: {
-        propName1: {}
-      }
-    });
-
-    var el1 = elem.create();
-    var el2 = elem2.create();
-
-    el1.appendChild(el2);
-    helperFixture().appendChild(el1);
-
-    // We need to check the number of calls as well which element triggered the
-    // property event.
-    el1.addEventListener('skate.property', () => ++triggeredNumber);
-    el1.addEventListener('skate.property', e => triggeredTarget = e.target);
-
-    // Will notify el1.propName1.
-    el2.propName1 = true;
-
-    // One for the property, and two for the event triggered once by the
-    // descendant and once by the host.
-    expect(triggeredNumber).to.equal(3);
-    expect(triggeredTarget).to.equal(el1);
-  });
-
-  it('attribute triggers property change', function () {
-    var triggered;
-
-    skate(elem.safe, {
-      properties: {
-        propName1: {
-          attr: true,
-          type: Boolean
-        }
-      }
-    });
-
-    var el = elem.create();
-    el.addEventListener('skate.property', e => triggered = e.detail.newValue);
-    el.setAttribute('prop-name1', '');
-    expect(triggered).to.equal(true);
-  });
-
   it('default value (scalar)', function () {
     skate(elem.safe, {
       properties: {
         propName1: {
-          value: 'test'
+          init: 'test'
         }
       }
     });
@@ -233,7 +133,7 @@ describe('api/property', function () {
     skate(elem.safe, {
       properties: {
         propName1: {
-          value: function () {
+          init: function () {
             return 'test';
           }
         }
@@ -244,29 +144,12 @@ describe('api/property', function () {
     expect(el.propName1).to.equal('test');
   });
 
-  it('should not notify if set to a falsy value', function () {
-    var triggered = false;
-
-    skate(elem.safe, {
-      properties: {
-        prop: {
-          notify: false
-        }
-      }
-    });
-
-    var el = elem.create();
-    el.addEventListener('skate.property', () => triggered = true);
-    el.prop = true;
-    expect(triggered).to.equal(false);
-  });
-
   it('initial value should not trump existing attribute values', function () {
     skate(elem.safe, {
       properties: {
         trump: {
           attr: true,
-          value: 'property'
+          init: 'property'
         }
       }
     });
